requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
certifi==2023.7.22
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.16.0
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.3.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
cryptography==41.0.5
    # via
    #   -r requirements/base.txt
    #   pyjwt
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
django==3.2.22
    # via
    #   -r requirements/base.txt
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   social-auth-app-django
django-cors-headers==4.3.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-extensions==3.2.3
    # via -r requirements/base.txt
django-rest-swagger==2.2.0
    # via -r requirements/base.txt
django-simple-history==3.0.0
    # via -r requirements/base.txt
django-waffle==4.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   drf-jwt
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-auth-backends==4.2.0
    # via -r requirements/base.txt
edx-django-release-util==1.3.0
    # via -r requirements/base.txt
edx-django-utils==5.7.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.13.0
    # via -r requirements/base.txt
edx-opaque-keys==2.5.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-rest-api-client==5.6.1
    # via -r requirements/base.txt
gevent==23.9.1
    # via -r requirements/production.in
greenlet==3.0.1
    # via gevent
gunicorn==21.2.0
    # via -r requirements/production.in
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   coreschema
markupsafe==2.1.3
    # via
    #   -r requirements/base.txt
    #   jinja2
mysqlclient==2.2.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/production.in
newrelic==9.1.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
packaging==23.2
    # via gunicorn
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
psutil==5.9.6
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pyjwt[crypto]==2.8.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwt
    #   social-auth-core
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
python-memcached==1.59
    # via -r requirements/production.in
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2023.3.post1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   -r requirements/production.in
    #   edx-django-release-util
    #   responses
requests==2.31.0
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   requests-oauthlib
    #   responses
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
responses==0.23.3
    # via -r requirements/base.txt
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
simplejson==3.19.2
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   edx-django-release-util
    #   python-memcached
slumber==0.7.1
    # via
    #   -r requirements/base.txt
    #   edx-rest-api-client
social-auth-app-django==5.4.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
social-auth-core==4.5.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-opaque-keys
types-pyyaml==6.0.12.12
    # via
    #   -r requirements/base.txt
    #   responses
typing-extensions==4.8.0
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   edx-opaque-keys
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   coreapi
urllib3==2.0.7
    # via
    #   -r requirements/base.txt
    #   requests
    #   responses
zope-event==5.0
    # via gevent
zope-interface==6.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
